requirements.txt
selenium==4.25.0
pytest==8.3.3
webdriver-manager==4.0.2
pytest-html
allure-pytest
